+++ src/main/java/fabrica/gestiondeproducciones/dominio/AnalisisYogur.java
@@ -2,28 +2,28 @@
 package fabrica.gestiondeproducciones.dominio;
 
 public class AnalisisYogur extends Analisis{
-    private int ph;
+    private float ph;
     private Produccion produccion;
 
     public AnalisisYogur() {
     }
 
-    public AnalisisYogur(int ph, Produccion produccion) {
+    public AnalisisYogur(float ph, Produccion produccion) {
         this.ph = ph;
         this.produccion = produccion;
     }
 
-    public AnalisisYogur(int ph, Produccion produccion, int id, Empleado encargado, String fecha, int levadura, int mos, int poliformosTotales, int poliformosFecales, String tipo) {
+    public AnalisisYogur(float ph, Produccion produccion, int id, Empleado encargado, String fecha, int levadura, int mos, int poliformosTotales, int poliformosFecales, String tipo) {
         super(id, encargado, fecha, levadura, mos, poliformosTotales, poliformosFecales, tipo);
         this.ph = ph;
         this.produccion = produccion;
     }
 
-    public int getPh() {
+    public float getPh() {
         return ph;
     }
 
-    public void setPh(int ph) {
+    public void setPh(float ph) {
         this.ph = ph;
     }
